move diff emission logic into submodule
